Implemented the new "global" type inference algorithm, only for global variables. Part of #2390

+++ spec/compiler/type_inference/class_spec.cr
@@ -733,7 +733,7 @@ describe "Type inference: class" do
           1
         end
 
-        $x = self.foo
+        $x = self.foo as Int32
       end
 
       $x
